+++ packages/amplify-category-api/src/commands/api/import.ts
@@ -6,12 +6,14 @@ import { RDS_SCHEMA_FILE_NAME } from '@aws-amplify/graphql-transformer-core';
 import { importAppSyncAPIWalkthrough } from '../../provider-utils/awscloudformation/service-walkthroughs/import-appsync-api-walkthrough';
 import { getAPIResourceDir } from '../../provider-utils/awscloudformation/utils/amplify-meta-utils';
 import { writeSchemaFile, generateRDSSchema } from '../../provider-utils/awscloudformation/utils/graphql-schema-utils';
+import { PREVIEW_BANNER } from '../../category-constants';
 
 const subcommand = 'import';
 
 export const name = subcommand;
 
 export const run = async (context: $TSContext) => {
+  printer.warn(PREVIEW_BANNER);
   const importAppSyncAPIWalkInputs = await importAppSyncAPIWalkthrough(context);
 
   if (importAppSyncAPIWalkInputs?.dataSourceConfig) {