+++ packages/amplify-category-api/src/commands/api/update-secrets.ts
@@ -8,16 +8,18 @@ import { databaseConfigurationInputWalkthrough } from '../../provider-utils/awsc
 import { getAppSyncAPIName, getAPIResourceDir } from '../../provider-utils/awscloudformation/utils/amplify-meta-utils';
 import {
   storeConnectionSecrets,
-  testDatabaseConnection,
   getSecretsKey,
   getDatabaseName,
-} from '../../provider-utils/awscloudformation/utils/rds-secrets/database-secrets';
+} from '../../provider-utils/awscloudformation/utils/rds-resources/database-resources';
+import { PREVIEW_BANNER } from '../../category-constants';
 
 const subcommand = 'update-secrets';
 
 export const name = subcommand;
 
 export const run = async (context: $TSContext) => {
+  printer.warn(PREVIEW_BANNER);
+
   const apiName = getAppSyncAPIName();
   const apiResourceDir = getAPIResourceDir(apiName);
 
@@ -30,13 +32,11 @@ export const run = async (context: $TSContext) => {
 
   const engine = ImportedRDSType.MYSQL;
   const secretsKey = await getSecretsKey();
-  const database = await getDatabaseName(context, apiName, secretsKey);
 
   // read and validate the RDS connection parameters
-  const databaseConfig: ImportedDataSourceConfig = await databaseConfigurationInputWalkthrough(engine, database);
+  const databaseConfig: ImportedDataSourceConfig = await databaseConfigurationInputWalkthrough(engine);
 
-  await testDatabaseConnection(databaseConfig);
   await storeConnectionSecrets(context, databaseConfig, apiName, secretsKey);
 
-  printer.info(`Successfully updated the secrets for ${database} database.`);
+  printer.info('Successfully updated the secrets for the database.');
 };