+++ packages/amplify-category-api/src/index.ts
@@ -24,12 +24,13 @@ import { authConfigToAppSyncAuthType } from './provider-utils/awscloudformation/
 import { checkAppsyncApiResourceMigration } from './provider-utils/awscloudformation/utils/check-appsync-api-migration';
 import { getAppSyncApiResourceName } from './provider-utils/awscloudformation/utils/getAppSyncApiName';
 import { getAPIResourceDir } from './provider-utils/awscloudformation/utils/amplify-meta-utils';
-import { configureMultiEnvDBSecrets } from './provider-utils/awscloudformation/utils/rds-secrets/multi-env-database-secrets';
+import { configureMultiEnvDBSecrets } from './provider-utils/awscloudformation/utils/rds-resources/multi-env-database-secrets';
 import {
   deleteConnectionSecrets,
   getSecretsKey,
   getDatabaseName,
-} from './provider-utils/awscloudformation/utils/rds-secrets/database-secrets';
+  removeVpcSchemaInspectorLambda,
+} from './provider-utils/awscloudformation/utils/rds-resources/database-resources';
 import { AmplifyGraphQLTransformerErrorConverter } from './errors/amplify-error-converter';
 
 export { NETWORK_STACK_LOGICAL_ID } from './category-constants';
@@ -306,6 +307,7 @@ export const handleAmplifyEvent = async (context: $TSContext, args: any): Promis
         return;
       }
       await deleteConnectionSecrets(context, apiName, args?.data?.envName);
+      await removeVpcSchemaInspectorLambda(context);
       break;
     default:
     // other event handlers not implemented