+++ packages/amplify-graphql-transformer-core/src/transformation/transform.ts
@@ -5,6 +5,8 @@ import {
   TransformHostProvider,
   TransformerLog,
   NestedStackProvider,
+  VpcConfig,
+  RDSLayerMapping,
 } from '@aws-amplify/graphql-transformer-interfaces';
 import type { AssetProvider, TransformParameters } from '@aws-amplify/graphql-transformer-interfaces';
 import { AuthorizationMode, AuthorizationType } from 'aws-cdk-lib/aws-appsync';
@@ -80,6 +82,8 @@ export interface GraphQLTransformOptions {
   readonly host?: TransformHostProvider;
   readonly userDefinedSlots?: Record<string, UserDefinedSlot[]>;
   readonly resolverConfig?: ResolverConfig;
+  readonly sqlLambdaVpcConfig?: VpcConfig;
+  readonly rdsLayerMapping?: RDSLayerMapping;
 }
 
 export type TransformOption = {
@@ -103,7 +107,9 @@ export class GraphQLTransform {
 
   private readonly userDefinedSlots: Record<string, UserDefinedSlot[]>;
 
+  private readonly sqlLambdaVpcConfig?: VpcConfig;
   private readonly transformParameters: TransformParameters;
+  private readonly rdsLayerMapping?: RDSLayerMapping;
 
   // A map from `${directive}.${typename}.${fieldName?}`: true
   // that specifies we have run already run a directive at a given location.
@@ -135,6 +141,8 @@ export class GraphQLTransform {
     this.stackMappingOverrides = options.stackMapping || {};
     this.userDefinedSlots = options.userDefinedSlots || ({} as Record<string, UserDefinedSlot[]>);
     this.resolverConfig = options.resolverConfig || {};
+    this.sqlLambdaVpcConfig = options.sqlLambdaVpcConfig;
+    this.sqlLambdaVpcConfig = options.sqlLambdaVpcConfig;
     this.transformParameters = {
       ...defaultTransformParameters,
       ...(options.transformParameters ?? {}),
@@ -191,6 +199,8 @@ export class GraphQLTransform {
       this.transformParameters,
       this.resolverConfig,
       datasourceConfig?.datasourceSecretParameterLocations,
+      this.sqlLambdaVpcConfig,
+      this.rdsLayerMapping,
     );
     const validDirectiveNameMap = this.transformers.reduce(
       (acc: any, t: TransformerPluginProvider) => ({ ...acc, [t.directive.name.value]: true }),